backend/requirements.txt
fastapi
uvicorn
sentence-transformers
umap-learn
scikit-learn
numpy
pydantic
